pyproject.toml
[tool.poetry]
name = "paperpile-notion"
version = "0.2.0"
description = "Sync Notion with Paperpile"
authors = ["J Muchovej <5000729+jmuchovej@users.noreply.github.com>"]
license = "MIT"
include = ["docs/config.yaml"]
readme = "README.md"
repository = "https://github.com/jmuchovej/paperpile-notion"

[tool.poetry.scripts]
paperpile-notion = "paperpile_notion.commands:cli"

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
"ruamel.yaml" = "^0.17.13"
click = "^8.0.1"
emojis = "^0.6.0"
bibtexparser = "^1.2.0"
notion-client = "^0.6.0"
attrs = "^21.2.0"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
pydocstyle = "^6.1.1"
pycodestyle = "^2.7.0"
pylama = "^7.7.1"
autopep8 = "^1.5.7"
ipython = "^7.27.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 88
ignore = "E501"
in-place = true
recursive = true
aggressive = 3
